Move logging into its own file for importing across modules
